refactoring (src folder)
