+++ run.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python3.6
+
+from src import main
+
+main.init()
