+++ src/main.py
@@ -1,15 +1,13 @@
-#!/usr/bin/env python3.6
+"the main"
 
 
 import os
 import signal
-import const
-import finder
-import system
 
-
-from table import *
-from dormitory import *
+from src import row as r
+from src import table as t
+from src import dormitory as d
+from src import const, finder, system
 
 
 def print_help():
@@ -20,6 +18,8 @@ def print_help():
 def init():
     "init function"
 
+    signal.signal(signal.SIGINT, system.signal_handler)
+
     if (len(system.argv) == 1) or (len(system.argv) == 2 and (system.argv[1] == '-h' or system.argv[1] == '--help')):
         print_help()
         system.exit(0)
@@ -27,29 +27,20 @@ def init():
     elif len(system.argv) > 6:
         system.error('Wrong arguments!\n', 1)
 
-    dorm = Dormitory(system.argv)
+    dorm = d.Dormitory(system.argv)
 
     name = "{} - {} floor".format(dorm.block().upper(), str(dorm.floor()))
-    table = Table(name)
+    table = t.Table(name)
 
     for number in dorm.rooms():
         names = finder.get(dorm.block(), number)
         if names is None:
             if dorm.empty_rooms():
-                table.add_row(Row(number)) # empty room
+                table.add_row(r.Row(number)) # empty room
             else:
                 continue
         else:
             for person in names:
-                table.add_row(Row(number,person))
+                table.add_row(r.Row(number,person))
 
     print(table)
-
-
-###############################################################################
-
-
-if __name__ == "__main__":
-    signal.signal(signal.SIGINT, system.signal_handler)
-    init()
-
