+++ src/table.py
@@ -1,7 +1,7 @@
 "output table"
 
 
-from row import *
+from src import const
 from itertools import groupby
